# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Trading Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /strategies:
    get:
      operationId: listStrategies
      tags: [strategies]
      summary: List all trading strategies
      responses:
        "200":
          description: List of strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
    post:
      operationId: createStrategy
      tags: [strategies]
      summary: Create a new trading strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStrategyBody"
      responses:
        "201":
          description: Created strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}:
    get:
      operationId: getStrategy
      tags: [strategies]
      summary: Get a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
        "404":
          description: Not found
    patch:
      operationId: updateStrategy
      tags: [strategies]
      summary: Update a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStrategyBody"
      responses:
        "200":
          description: Updated strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
    delete:
      operationId: deleteStrategy
      tags: [strategies]
      summary: Delete a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Bot Control ───────────────────────────────────────────────────
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List trade history
      parameters:
        - name: brokerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: strategyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: symbol
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a single trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found

  # ── Reports ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Strategy:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        brokerId:
          type: integer
        assetType:
          type: string
          description: "stocks | options | crypto | etf"
        symbols:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        maxPositionSize:
          type: number
          description: "Max $ per position"
        maxDailyLoss:
          type: number
          description: "Max daily loss in $"
        stopLossPercent:
          type: number
          description: "Stop loss as % from entry"
        takeProfitPercent:
          type: number
          description: "Take profit as % from entry"
        aiEnabled:
          type: boolean
        aiModel:
          type: ["string", "null"]
        aiSignalThreshold:
          type: number
          description: "Minimum AI confidence 0-1"
        rsiOverbought:
          type: ["number", "null"]
        rsiOversold:
          type: ["number", "null"]
        maFastPeriod:
          type: ["integer", "null"]
        maSlowPeriod:
          type: ["integer", "null"]
        tradeCount:
          type: integer
        winRate:
          type: ["number", "null"]
        totalPnl:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, brokerId, assetType, symbols, isActive, maxPositionSize, maxDailyLoss, stopLossPercent, takeProfitPercent, aiEnabled, aiSignalThreshold, tradeCount, createdAt, updatedAt]
    CreateStrategyBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        brokerId:
          type: integer
        assetType:
          type: string
        symbols:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        maxPositionSize:
          type: number
        maxDailyLoss:
          type: number
        stopLossPercent:
          type: number
        takeProfitPercent:
          type: number
        aiEnabled:
          type: boolean
        aiModel:
          type: ["string", "null"]
        aiSignalThreshold:
          type: number
        rsiOverbought:
          type: ["number", "null"]
        rsiOversold:
          type: ["number", "null"]
        maFastPeriod:
          type: ["integer", "null"]
        maSlowPeriod:
          type: ["integer", "null"]
      required: [name, brokerId, assetType, symbols, maxPositionSize, maxDailyLoss, stopLossPercent, takeProfitPercent, aiSignalThreshold]
    UpdateStrategyBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        symbols:
          type: ["array", "null"]
          items:
            type: string
        isActive:
          type: ["boolean", "null"]
        maxPositionSize:
          type: ["number", "null"]
        maxDailyLoss:
          type: ["number", "null"]
        stopLossPercent:
          type: ["number", "null"]
        takeProfitPercent:
          type: ["number", "null"]
        aiEnabled:
          type: ["boolean", "null"]
        aiSignalThreshold:
          type: ["number", "null"]
        rsiOverbought:
          type: ["number", "null"]
        rsiOversold:
          type: ["number", "null"]
        maFastPeriod:
          type: ["integer", "null"]
        maSlowPeriod:
          type: ["integer", "null"]

    # ── Bot ─────────────────────────────────────────────────────────
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        brokerId:
          type: integer
        brokerName:
          type: string
        strategyId:
          type: ["integer", "null"]
        strategyName:
          type: ["string", "null"]
        symbol:
          type: string
        assetType:
          type: string
        side:
          type: string
          description: "buy | sell | buy_to_open | sell_to_close"
        quantity:
          type: number
        entryPrice:
          type: number
        exitPrice:
          type: ["number", "null"]
        realizedPnl:
          type: ["number", "null"]
        realizedPnlPercent:
          type: ["number", "null"]
        status:
          type: string
          description: "open | closed | cancelled"
        aiSignal:
          type: ["string", "null"]
          description: "buy | sell | hold"
        aiConfidence:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        openedAt:
          type: string
        closedAt:
          type: ["string", "null"]
      required: [id, brokerId, brokerName, symbol, assetType, side, quantity, entryPrice, status, openedAt]

    # ── Reports ─────────────────────────────────────────────────────
...